chore(): test case skeleton

+++ remove.js
@@ -7,8 +7,16 @@ async function remove({
   did = '',
   password = ''
 } = {}) {
+
+  if (0 === paths.length) {
+    throw new Error("No path(s) provided.")
+  }
+
   const { afs } = await create({ did, password })
   for (const path of paths) {
+    if ('string' !== typeof path) {
+      throw new Error("Path found that is not of type string", path)
+    }
     try {
       if (await afs.access(path)) {
         await afs.unlink(path)
